Move the `wcdl/subsection` directory into the `components` directory and rename related CSS.
