+++ js/src/components/account-card/index.js
@@ -9,7 +9,7 @@ import { Icon, store as storeIcon } from '@wordpress/icons';
  * Internal dependencies
  */
 import Section from '~/components/section';
-import Subsection from '~/wcdl/subsection';
+import Subsection from '~/components/subsection';
 import googleLogoURL from '~/images/logo/gogole-g-logo.svg';
 import googleMCLogoURL from '~/images/logo/google-merchant-center-logo.svg';
 import googleAdsLogoURL from '~/images/logo/google-ads-logo.svg';
