+++ js/src/components/contact-information/contact-information-preview-card.scss
@@ -1,6 +1,6 @@
 .gla-contact-info-preview-card {
 	// Vertically align icon inside the title.
-	.wcdl-subsection-title {
+	.gla-subsection-title {
 		display: flex;
 		align-items: center;
 	}
