+++ js/src/components/free-listings/choose-audience-section/choose-audience-section.scss
@@ -3,7 +3,7 @@
 		font-style: normal;
 	}
 
-	.wcdl-subsection-helper-text {
+	.gla-subsection-helper-text {
 		margin-bottom: calc(var(--main-gap) / 3 * 2);
 	}
