+++ js/src/components/google-ads-account-card/connect-ads/index.scss
@@ -3,7 +3,7 @@
 		flex-grow: 1;
 	}
 
-	.wcdl-subsection-body {
+	.gla-subsection-body {
 		margin-bottom: $grid-unit;
 	}
 }
