+++ js/src/components/google-mc-account-card/reclaim-url-card/index.js
@@ -15,7 +15,7 @@ import { getSetting } from '@woocommerce/settings'; // eslint-disable-line impor
 import AppDocumentationLink from '~/components/app-documentation-link';
 import AppButton from '~/components/app-button';
 import Section from '~/components/section';
-import Subsection from '~/wcdl/subsection';
+import Subsection from '~/components/subsection';
 import useApiFetchCallback from '~/hooks/useApiFetchCallback';
 import useExistingGoogleMCAccounts from '~/hooks/useExistingGoogleMCAccounts';
 import { useAppDispatch } from '~/data';
