+++ js/src/components/subsection/body.js
@@ -6,7 +6,7 @@ import './body.scss';
 const Body = ( props ) => {
 	const { children } = props;
 
-	return <div className="wcdl-subsection-body">{ children }</div>;
+	return <div className="gla-subsection-body">{ children }</div>;
 };
 
 export default Body;
