+++ js/src/components/subsection/body.scss
@@ -1,4 +1,4 @@
-.wcdl-subsection-body {
+.gla-subsection-body {
 	font-size: $default-font-size;
 	font-style: normal;
 	font-weight: 400;
