+++ js/src/components/subsection/helper-text.scss
@@ -1,4 +1,4 @@
-.wcdl-subsection-helper-text {
+.gla-subsection-helper-text {
 	font-size: $helptext-font-size;
 	font-style: italic;
 	font-weight: 400;
