+++ js/src/components/subsection/index.scss
@@ -1,4 +1,4 @@
-.wcdl-subsection {
+.gla-subsection {
 	&:not(:first-child) {
 		margin-top: var(--main-gap);
 	}
