+++ js/src/components/subsection/subtitle.scss
@@ -1,4 +1,4 @@
-.wcdl-subsection-subtitle {
+.gla-subsection-subtitle {
 	font-size: $gla-font-smaller;
 	line-height: $gla-line-height-smaller;
 	color: $gray-700;
