+++ js/src/components/subsection/title.scss
@@ -1,4 +1,4 @@
-.wcdl-subsection-title {
+.gla-subsection-title {
 	position: relative;
 	font-size: $gla-font-small;
 	font-style: normal;
