+++ js/src/pages/settings/index.scss
@@ -15,7 +15,7 @@
 		}
 	}
 
-	.wcdl-subsection-helper-text {
+	.gla-subsection-helper-text {
 		margin-top: $grid-unit-10;
 		color: $gray-600;
 	}
